+++ examples/llvm/at.c
@@ -42,6 +42,7 @@
 #endif
 
 #include <pthread.h>
+#include <stdint.h>
 
 // For native execution.
 #ifndef DIVINE
@@ -63,7 +64,7 @@ LTL(exclusion, G(!(critical1 && critical2)));
 #define DELTA  1
 
 int *timer;
-int x, y = 0, z = 0;
+intptr_t x, y = 0, z = 0;
 int finished = 0;
 
 int _critical = 0;
@@ -76,7 +77,7 @@ void critical() {
 }
 
 void *fnc_thread( void *arg ) {
-    int id = (int) arg;
+    intptr_t id = ( intptr_t ) arg;
 
     if ( id == 1 )
         ap( wait1 );
@@ -167,7 +168,7 @@ int main() {
     pthread_create( &threads[0], 0, fnc_timer, NULL );
     for ( i=1; i <= NUM_OF_THREADS; i++ ) {
         timer[i-1] = OFF;
-        pthread_create( &threads[i], 0, fnc_thread, ( void* )( i ) );
+        pthread_create( &threads[i], 0, fnc_thread, ( void* )( intptr_t )( i ) );
     }
 
     for ( i=1; i <= NUM_OF_THREADS; i++ ) {
